requirements.txt
apache-airflow
numpy
pandas
dlt
openpyxl
unidecode
sqlalchemy
psycopg2-binary
dbt-core
dbt-postgres